+++ Howatworks.PlayerJournal.Serialization/Combat/CapShipBond.cs
@@ -3,7 +3,7 @@
     // TODO: no sample
     public class CapShipBond : JournalEntryBase
     {
-        public int Reward { get; set; }
+        public long Reward { get; set; }
         public string AwardingFaction { get; set; }
         public string VictimFaction { get; set; }
     }
