+++ Howatworks.PlayerJournal.Serialization/Exploration/SellExplorationData.cs
@@ -6,8 +6,8 @@ namespace Howatworks.PlayerJournal.Serialization.Exploration
     {
         public List<string> Systems { get; set; }
         public List<string> Discovered { get; set; }
-        public int BaseValue { get; set; }
-        public int Bonus { get; set; }
-        public int TotalEarnings { get; set; }
+        public long BaseValue { get; set; }
+        public long Bonus { get; set; }
+        public long TotalEarnings { get; set; }
     }
 }
