+++ Howatworks.PlayerJournal.Serialization/Trade/BuyTradeData.cs
@@ -3,6 +3,6 @@
     public class BuyTradeData : JournalEntryBase
     {
         public string System { get; set; } // Note: name
-        public int Cost { get; set; }
+        public long Cost { get; set; }
     }
 }
