+++ Howatworks.PlayerJournal.Serialization/Trade/MarketBuy.cs
@@ -8,7 +8,7 @@ namespace Howatworks.PlayerJournal.Serialization.Trade
         public long MarketID { get; set; }
         public string Type { get; set; }
         public int Count { get; set; }
-        public int BuyPrice { get; set; }
-        public int TotalCost { get; set; }
+        public long BuyPrice { get; set; }
+        public long TotalCost { get; set; }
     }
 }
